requirements.txt
agno>=0.2.0
python-dotenv>=1.0.0
anthropic>=0.34.0
